:truck: rename: page 및 component 폴더 구조 잡기
