+++ src/components/home/pages/index.tsx
@@ -1,0 +1,5 @@
+function Home() {
+  return <h1>Home page</h1>;
+}
+
+export default Home;
